fix: add loader on fly
