+++ components/map/controls/geolocation-control/geolocation-control.module.css
@@ -1,0 +1,3 @@
+.geolocationControl:disabled div:has(> svg) {
+  margin: 0;
+}
